requirements.txt
pandas==1.4.4
pillow==9.2.0
streamlit==1.12.2
geopandas==0.11.1
shapely==1.8.4
matplotlib==3.5.3
mapclassify==2.4.3
geopy==2.2.0
pyproj==3.4.0
rtree==1.0.0
openpyxl==3.0.10